+++ src/GlobalPayments.Api/Entities/GpApi/GpApiAuthorizationRequestBuilder.cs
@@ -25,8 +25,8 @@ namespace GlobalPayments.Api.Entities {
                 .Set("avs_address", builder.BillingAddress?.StreetAddress1)
                 .Set("avs_postal_code", builder.BillingAddress?.PostalCode)
                 .Set("funding", builder.PaymentMethod?.PaymentMethodType == PaymentMethodType.Debit ? "DEBIT" : "CREDIT") // [DEBIT, CREDIT]
-                .Set("authcode", builder.OfflineAuthCode);
-                //.Set("brand_reference", "")
+                .Set("authcode", builder.OfflineAuthCode)
+                .Set("brand_reference", builder.CardBrandTransactionId);
 
                 card.Set("chip_condition", EnumConverter.GetMapping(Target.GP_API, builder.EmvChipCondition)); // [PREV_SUCCESS, PREV_FAILED]