+++ src/GlobalPayments.Api/ServiceConfigs/Gateways/GpApiConfig.cs
@@ -50,6 +50,10 @@ namespace GlobalPayments.Api {
         /// 3DSecure method return url
         /// </summary>
         public string MethodNotificationUrl { get; set; }
+        /// <summary>
+        /// MerchantContactUrl
+        /// </summary>
+        public string MerchantContactUrl { get; set; }
 
         public GpApiConfig() : base(GatewayProvider.GP_API) { }
 
@@ -80,6 +84,7 @@ namespace GlobalPayments.Api {
                 ChallengeNotificationUrl = ChallengeNotificationUrl,
                 MethodNotificationUrl = MethodNotificationUrl,
                 RequestLogger = RequestLogger,
+                MerchantContactUrl = MerchantContactUrl,
                 WebProxy = WebProxy,
             };