+++ tests/GlobalPayments.Api.Tests/GpApi/GpApi3DSecure2Tests.cs
@@ -2,6 +2,7 @@
 using GlobalPayments.Api.Entities;
 using GlobalPayments.Api.PaymentMethods;
 using GlobalPayments.Api.Services;
+using GlobalPayments.Api.Tests.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace GlobalPayments.Api.Tests.GpApi {
@@ -31,6 +32,8 @@ namespace GlobalPayments.Api.Tests.GpApi {
                 Country = "GB",
                 ChallengeNotificationUrl = "https://ensi808o85za.x.pipedream.net/",
                 MethodNotificationUrl = "https://ensi808o85za.x.pipedream.net/",
+                MerchantContactUrl = "https://enp4qhvjseljg.x.pipedream.net/",
+                RequestLogger = new RequestFileLogger(@"C:\temp\gpapi\requestlog.txt")
             });
         }